refactor out the markdown link parsers into a separate file
